requirements.txt
Django==2.1.4
djangorestframework==3.9.0
flake8==3.6.0
Markdown==3.0.1
